Generator fixes + interface for marker type
Can now generate code that handles multiple "ret" statements in the
try/catch. Marker type should now either be a class called
MethodDecoratorAttribute (or derived from it) or an interface called
IMethodDecorator. This allows for the implementation of the attribute to
be in a base class in a separate assembly